ref: extract typedData logic to utility function in uncontrolled form
